requirements.txt
-i https://pypi.org/simple
aiofiles==22.1.0; python_version >= '3.7' and python_version < '4.0'
dnspython==2.2.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
httptools==0.6.4; python_full_version >= '3.5.0'
jinja2==3.1.2
markupsafe==2.1.1; python_version >= '3.7'
motor==3.7.0
multidict==6.0.5; python_version >= '3.7'
natural==0.2.0
pyjwt==2.10.1
pymongo==4.11.2
python-dateutil==2.8.2
python-dotenv==0.18.0
requests==2.32.3
sanic==24.12.0
sanic-routing==23.12.0
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ujson==5.5.0; sys_platform != 'win32' and implementation_name == 'cpython'
uvloop==0.17.0; sys_platform != 'win32' and implementation_name == 'cpython'
websockets==10.3; python_version >= '3.7'
